refactor vapi routes into modules
